// src/store/modules/jt-cesium-vue/modules/layout/mutation-types.ts
export enum LayoutMutationTypes {
  RESET_STATE = 'RESET_STATE',
  SET_SHOW_BROWSER_PANEL = 'SET_SHOW_BROWSER_PANEL',
  SET_SHOW_TOOLBAR = 'SET_SHOW_TOOLBAR',
  SET_SHOW_SETTING_BUTTON = 'SET_SHOW_SETTING_BUTTON',
  SET_TOOLBAR_HEIGHT = 'SET_TOOLBAR_HEIGHT',

  ADD_UNIQUE_NAME_OVERLAY_DYNAMIC_VIEW_BY_NAME = 'ADD_UNIQUE_NAME_OVERLAY_DYNAMIC_VIEW_BY_NAME',
  REMOVE_OVERLAY_DYNAMIC_VIEW_BY_NAME = 'REMOVE_OVERLAY_DYNAMIC_VIEW_BY_NAME',
}