// src/store/modules/jt-cesium-vue/modules/layout/action-types.ts
export enum LayoutActionTypes {
  RESET_STATE = 'RESET_STATE',
  SET_SHOW_BROWSER_PANEL = 'SET_SHOW_BROWSER_PANEL',
  SET_SHOW_TOOLBAR = 'SET_SHOW_TOOLBAR',
  SET_TOOLBAR_HEIGHT = 'SET_TOOLBAR_HEIGHT',

  ADD_UNIQUE_NAME_OVERLAY_DYNAMIC_VIEW_BY_NAME = 'ADD_UNIQUE_NAME_OVERLAY_DYNAMIC_VIEW_BY_NAME',
  REMOVE_OVERLAY_DYNAMIC_VIEW_BY_NAME = 'REMOVE_OVERLAY_DYNAMIC_VIEW_BY_NAME',
}